requirements.txt
# Core dependencies
fastapi==0.95.0
uvicorn==0.22.0
gunicorn==20.1.0

# Database
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
alembic==1.9.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Scientific packages
numpy==1.20.3
pandas==1.3.0
openpyxl==3.0.9

# Other dependencies
python-dotenv==0.19.0
python-multipart==0.0.5
pydantic==1.10.7
email-validator==2.0.0

# Production monitoring
prometheus-fastapi-instrumentator==5.9.1
